Python-unittest/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
asgiref==3.5.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==36.0.2
ddt==1.4.4
Django==1.11a1
execnet==1.9.0
h11==0.13.0
HTMLTestRunner-Python3==0.8.0
idna==2.7
iniconfig==1.1.1
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
pytest==7.1.1
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
pyttest==0.1
pytz==2021.3
reload==0.9
requests==2.20.0
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.24.3
watchdog==2.1.7
wsproto==1.1.0
xlrd==1.2.0
